README
Do I regret anything I have done in my life, the answer no. Would I change myself to fit other people idles, on how they see thing or the way they make you change the way you see yourself or the world you live in just so you can have a bit of happiness for only the second or seconds that they choose to give you in there time line of life. If time is the best thing you can give someone then it doesn't matter how much time you get or doses it.
Because with in that second you are happy and content with a person it could be but a second but feel like a lifetime. But is this the case with all so why do you choose to give more to others than to the ones you should give it to. To the ones that are your true friends and family not the ones that are lost or time wasters. The ones that will take and take as much as they can until you have nothing left to give them or other or even to yourself. 
Then why do you alow this to keep happening to you or the good ones around you time and again. Is it with in everyone to suck out your live and the true you from within yourself just so you can fit into there ideas and posepsons of there world they live in. Or could it just be that you don't really fit in and are looking for some justification of your own reasons for being here and now. 
I have never understood the concept of being in the moment I can't seem to get my head around that. Is it to have a second taken for your self to understand what is going on around you to take note of your environment, to take note or the people around you or to just take the time to relax and enjoy the time you have been given to be able to take that second or seconds for yourself to clocked your self and dust yourself of and try ones again this this great journey to discover or true part in our own insurance life. To grow as we get older and to learn to just be in at pice with the past and not to get to  upset with the way things in our life go?
Is it to find the joy with in that moment so you have something to hold on to next time you get through a big pile of carp into your lap. Dose it make you stronger when the chips are stacked against you. So you can look back on the happy times the times you wish you still had and the wasted times you swindle to find that single moment of happiness with in yourself. Is this the way is meant to be or is it the way that the ones who have lead you astray leave on you like a contest remember of the space and time they once filled in your life?
The silence of the hidden screen and the invisible scars we all have and hide from the world and from the ones around as. The pains from the past that try as hard as we can we are unable to escape from, the dreams in the night when all is quiet and should be a time of rest to renew yourself for the next coming day, just remembered you of the things you had no control over or the things you chooses to not take notes off. Some night peaceful sleep is given with no dreams of pasted events and other night where it is as real as the days or day you lived through it a crazy nights where you move back in time to understand and to be able to recover or reflect on the past to grow from and to analyse your own misstates. To be able to grow again not as the one how lived through it but as someone how can speak cool cram and with no emotions behind your past events. As in a way been removed not attached to the pain or regular you once held with in that moment in your life.
I am a big believer that from my past will help make me the porsion I will become in the future but my past will not define me. You mite sit here thinking WTF but is really quite simple. Just because shit has happened it doesn't matter you can't change the past all you can do is grow from it get smart try hard don't forget to uses your own gut feelings on thing so in turn you don't find yourself down the track in like at the same round about. Just because I did something wrong in my past dose not mean that I will do it again. Self growth.
We all are not perfect but it's the ones that understand they are not perfect that will work on themselves, to strive to be better with in themselves not for anyone or anything else but self understanding. It the ones that don't think this way I find myself not understanding and lost around. To push ones own fault and imperfections to someone as they are in a uttered perception an reality. Is the most lowest of things to do to someone in my opinion and really shows the truest of lows in the human nature. 
Then why? Do we as a whole allow this to keep happening, can we as a hole not stop it, or has it become so much of our day to day life that we are now blind bye it. That is the new normality or could it just be completely, and we have just become numbed to the world around as.
They say in the dubbhis region to gain everything you mast loses everything! But if you live with nothing you can live with everything!
My question is if you have nothing to loses, what if you have nothing to losses dose this mean you have nothing to gain.